Move game-components to it's own directory
