+++ docs/js/game/chat.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/game/chat.js -->
+
+## React
+
+<!-- End src/components/game/chat.js -->
+
