+++ docs/js/game/index.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/game/index.js -->
+
+## React
+
+<!-- End src/components/game/index.js -->
+
