+++ src/components/game/index.js
